requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
defusedxml==0.7.1
exceptiongroup==1.3.1
fastapi==0.127.1
idna==3.11
jira==3.10.5
oauthlib==3.3.1
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
